+++ src/Orbitones.cpp
@@ -3,17 +3,22 @@
 #define DISPLAY_SIZE_WIDTH 427
 #define DISPLAY_SIZE_HEIGHT 378
 #define MAX_PARTICLES 16
+#define MAX_HISTORY 20
 #define INTERNAL_SAMP_TIME 60.0
 
 struct Trail {
     float x;
     float y;
     int alpha;
-    Trail() {}
+    bool visible;
+    Trail() {
+        visible = false;
+    }
     Trail(float _x, float _y, int _alpha) {
         x = _x;
         y = _y;
         alpha = _alpha;
+        visible = true;
     }
 };
 
@@ -27,7 +32,10 @@ struct Particle {
     float radius;
     bool visible;
     bool whiteTrails = true;
-    std::vector<Trail> history;
+    // std::vector<Trail> history;
+    Trail history[MAX_HISTORY];
+    int currentHistorySize;
+    int historyIndex;
 
     Particle() {
         box.pos.x = randRange(0, DISPLAY_SIZE_WIDTH);
@@ -38,6 +46,8 @@ struct Particle {
         radius = randRange(5, 12);
         mass = radius;
         visible = false;
+        currentHistorySize = 0;
+        historyIndex = 0;
     }
 
     void setPos(Vec _pos) {
@@ -71,17 +81,22 @@ struct Particle {
         // trail
         Vec v = box.getCenter();
         Trail t = Trail(v.x, v.y, 255);
-        history.insert(history.begin(), t);
-        // history.push_back(t);
-        if (history.size() > 20) {
-            history.pop_back();
-            // history.erase(history.begin());
+
+        // shift elements to insert at beginning
+        for (int i = MAX_HISTORY-1; i > 0; i--) {
+            history[i] = history[i-1];
         }
+        history[0] = t;
+
+        currentHistorySize = std::min(++currentHistorySize, MAX_HISTORY);
+        historyIndex = (historyIndex + 1) % MAX_HISTORY;
+
         return true;
     }
 
     void clearHistory() {
-        history.clear();
+        currentHistorySize = 0;
+        historyIndex = 0;
     }
 };
 
@@ -138,6 +153,7 @@ struct Orbitones : Module {
         NUM_ATTRACTORS
     };
     enum ParamIds {
+        OFFSET_PARAM,
         REMOVE_PARTICLE_PARAM,
         CLEAR_PARTICLES_PARAM,
         MOVE_ATTRACTORS_PARAM,
@@ -173,6 +189,7 @@ struct Orbitones : Module {
     dsp::SchmittTrigger removeTrig, clearTrig, moveTrig;
     Attractor *attractors = new Attractor[NUM_ATTRACTORS];
     Particle *particles = new Particle[MAX_PARTICLES];
+    float voltsOffset = 0.0;
     bool movement = false;
     bool drawTrails = true;
     bool particleBoundary = false;
@@ -184,6 +201,7 @@ struct Orbitones : Module {
 
     Orbitones() {
         config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
+        configParam(OFFSET_PARAM, -5.0, 5.0, 0.0, "Offset", " V");
         configParam(REMOVE_PARTICLE_PARAM, 0.0, 1.0, 0.0, "Remove previous particle");
         configParam(CLEAR_PARTICLES_PARAM, 0.0, 1.0, 0.0, "Clear particles");
         configParam(MOVE_ATTRACTORS_PARAM, 0.0, 1.0, 0.0, "Move attractors");
@@ -320,6 +338,7 @@ struct Orbitones : Module {
     }
 
     void process(const ProcessArgs &args) override {
+        voltsOffset = params[OFFSET_PARAM].getValue();
         if (removeTrig.process(params[REMOVE_PARTICLE_PARAM].getValue())) {
             if (visibleParticles > 0) removeParticle(visibleParticles-1);
         }
@@ -372,10 +391,11 @@ struct Orbitones : Module {
                         }
                     }
                     particles[i].update();
-                    float voltsX = rescale(particles[i].box.pos.x, 0, DISPLAY_SIZE_WIDTH, -5.0, 5.0);
-                    float voltsY = rescale(particles[i].box.pos.y, DISPLAY_SIZE_HEIGHT, 0, -5.0, 5.0);
-                    float voltsVelX = rescale(particles[i].vel.x, -12.0, 12.0, -5.0, 5.0);
-                    float voltsVelY = rescale(particles[i].vel.y, 12.0, -12.0, -5.0, 5.0);
+
+                    float voltsX = rescale(particles[i].box.pos.x, 0, DISPLAY_SIZE_WIDTH, -5.0 + voltsOffset, 5.0 + voltsOffset);
+                    float voltsY = rescale(particles[i].box.pos.y, DISPLAY_SIZE_HEIGHT, 0, -5.0 + voltsOffset, 5.0 + voltsOffset);
+                    float voltsVelX = rescale(particles[i].vel.x, -12.0, 12.0, -5.0 + voltsOffset, 5.0 + voltsOffset);
+                    float voltsVelY = rescale(particles[i].vel.y, 12.0, -12.0, -5.0 + voltsOffset, 5.0 + voltsOffset);
                     outputs[X_POLY_OUTPUT].setVoltage(voltsX, i);
                     outputs[Y_POLY_OUTPUT].setVoltage(voltsY, i);
                     outputs[NEG_X_POLY_OUTPUT].setVoltage(-voltsX, i);
@@ -694,8 +714,7 @@ struct OrbitonesDisplay : Widget {
                 if (module->drawTrails) {
                     bool updated = module->particles[i].updateHistory();
                     if (!updated) break;
-                    unsigned int length = std::min<unsigned int>(module->particles[i].history.size(), 20);
-                    for (unsigned int j = 1; j < length; j++) {
+                    for (int j = 1; j < module->particles[i].currentHistorySize; j++) {
                         Trail trailPos = module->particles[i].history[j];
                         Trail trailPosPrev = module->particles[i].history[j-1];
                         nvgBeginPath(args.vg);
@@ -746,6 +765,7 @@ struct OrbitonesWidget : ModuleWidget {
         addChild(createWidget<JeremyScrew>(Vec(505.5, 2)));
         addChild(createWidget<JeremyScrew>(Vec(505.5, box.size.y - 14)));
 
+        addParam(createParamCentered<TinyBlueKnob>(Vec(511.5, 201.5), module, Orbitones::OFFSET_PARAM));
         addParam(createParamCentered<TinyBlueButton>(Vec(42, 56), module, Orbitones::REMOVE_PARTICLE_PARAM));
         addParam(createParamCentered<TinyBlueButton>(Vec(42, 93.4), module, Orbitones::CLEAR_PARTICLES_PARAM));
         addParam(createParamCentered<TinyBlueButton>(Vec(42, 130.9), module, Orbitones::MOVE_ATTRACTORS_PARAM));