+++ src/plugin.cpp
@@ -12,6 +12,7 @@ void init(Plugin* p) {
 	p->addModel(modelStochSeq4);
 	p->addModel(modelPolyrhythmClock);
 	p->addModel(modelRandGates);
+	p->addModel(modelRandRoute);
 	p->addModel(modelNeutrinode);
 	p->addModel(modelCosmosis);
 	p->addModel(modelJeremyBlankPanel);
@@ -21,6 +22,7 @@ void init(Plugin* p) {
 	p->addModel(modelOrbitones);
 	p->addModel(modelAbsorptionSpectrum);
 	p->addModel(modelTalea);
+	p->addModel(modelCollider);
 
 	// Any other plugin initialization may go here.
 	// As an alternative, consider lazy-loading assets and lookup tables when your module is created to reduce startup times of Rack.